requirements.txt
faiss-cpu
gradio
langchain-community
langchain-openai
langchain-text-splitters
pypdf
python-dotenv
tiktoken
